+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 - Change the TimelineTranslations constructor to require all translations or use the TimelineTranslations.empty constructor if you don't want to specify all translations
 - Add a TimelinePaddingOptions class to store the padding options for the timeline
 - fix the avatar size to match the new design
+- Add the iconbutton for image uploading back to the ReactionBottom
 
 ## 3.0.1